renamed ticker.rb
