+++ lib/bitstampede/Mapper.rb
@@ -1,7 +1,7 @@
 require_relative './entities/UserBalance'
 require_relative './entities/Order'
 require_relative './entities/transaction'
-require_relative './entities/ticker'
+require_relative './entities/Ticker'
 
 module Bitstampede
   class Mapper
