+++ lib/bitstampede/entities/Ticker.rb
